// src/lib/api/generated/model/emailTemplate.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */
import type { TemplateTypeEnum } from './templateTypeEnum';
import type { EmailTemplateDesignJson } from './emailTemplateDesignJson';
import type { EmailTemplateStatusEnum } from './emailTemplateStatusEnum';

export interface EmailTemplate {
  readonly id: string;
  /** @nullable */
  readonly tenant: string | null;
  readonly created_by: string;
  /** @maxLength 255 */
  name: string;
  /**
   * @maxLength 100
   * @pattern ^[-a-zA-Z0-9_]+$
   */
  slug: string;
  template_type?: TemplateTypeEnum;
  /** @maxLength 255 */
  subject: string;
  html_content: string;
  plain_text_content?: string;
  required_variables?: unknown;
  /** @nullable */
  design_json?: EmailTemplateDesignJson;
  is_default?: boolean;
  /** @maxLength 10 */
  language?: string;
  status?: EmailTemplateStatusEnum;
  readonly created_at: string;
  readonly updated_at: string;
}